refactor update endpoint
