+++ backend/main.go
@@ -17,14 +17,14 @@ func main() {
 	dbName := os.Getenv("MYSQL_DATABASE")
 	dbmodel.DbConnectionString = fmt.Sprintf("%s:%s@tcp(%s:%s)/%s?charset=utf8mb4&parseTime=True&loc=Local", dbUser, dbPassword, dbHost, dbPort, dbName)
 
-	// connec to db + migrate db
+	// connect to db + migrate db
 	dbmodel.Connect(dbmodel.DbConnectionString)
 	dbmodel.InitialMigration(dbmodel.Db)
 
 	// insert articles
-	for i := 0; i < 5; i++ {
-		dbmodel.DbCreateArticle(dbmodel.Db, dbmodel.Article{Title: "Book Title", Desc: "Book Description", Content: "Book Content"})
-	}
+	// for i := 0; i < 5; i++ {
+	// 	dbmodel.DbCreateArticle(dbmodel.Db, dbmodel.Article{Title: "Book Title", Desc: "Book Description", Content: "Book Content"})
+	// }
 
 	// start web api
 	webserver.StartServer()
